+++ weblate/utils/search.py
@@ -156,7 +156,7 @@ class QueryParser(whoosh.qparser.QueryParser):
             "changed_by": TEXT,
             # Unit data
             "check": TEXT,
-            "ignored_check": TEXT,
+            "dismissed_check": TEXT,
             "suggestion": TEXT,
             "suggestion_author": TEXT,
             "comment": TEXT,
@@ -199,7 +199,7 @@ FIELD_MAP = {"changed": "change__timestamp", "added": "timestamp"}
 STRING_FIELD_MAP = {"suggestion": "suggestion__target", "comment": "comment__comment"}
 EXACT_FIELD_MAP = {
     "check": "check__check",
-    "ignored_check": "check__check",
+    "dismissed_check": "check__check",
     "language": "translation__language__code",
     "changed_by": "change__author__username",
     "suggestion_author": "suggestion__user__username",
@@ -229,9 +229,9 @@ def field_extra(field, query):
     if field in {"changed", "changed_by"}:
         return query & Q(change__action__in=Change.ACTIONS_CONTENT)
     if field == "check":
-        return query & Q(check__ignore=False)
-    if field == "ignored_check":
-        return query & Q(check__ignore=True)
+        return query & Q(check__dismissed=False)
+    if field == "dismissed_check":
+        return query & Q(check__dismissed=True)
     return query
 
 
@@ -259,9 +259,9 @@ def has_sql(text):
     if text in ("resolved-comment", "resolved_comment"):
         return Q(comment__resolved=True)
     if text in ("check", "failing-check", "failing_check"):
-        return Q(check__ignore=False)
-    if text in ("ignored-check", "ignored_check"):
-        return Q(check__ignore=True)
+        return Q(check__dismissed=False)
+    if text in ("dismissed-check", "dismissed_check", "ignored-check", "ignored_check"):
+        return Q(check__dismissed=True)
     if text == "translation":
         return Q(state__gte=STATE_TRANSLATED)
     if text == "shaping":